opt some

+++ etc/miss.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-require &>/dev/null || if [ -f $PWD/.ish/plug.sh ]; then source $PWD/.ish/plug.sh; elif [ -f $HOME/.ish/plug.sh ]; then source $HOME/.ish/plug.sh; else
+if [ -f $PWD/.ish/plug.sh ]; then source $PWD/.ish/plug.sh; elif [ -f $HOME/.ish/plug.sh ]; then source $HOME/.ish/plug.sh; else
 	ctx_temp=$(mktemp); if curl -h &>/dev/null; then curl -o $ctx_temp -fsSL https://shylinux.com; else wget -O $ctx_temp -q http://shylinux.com; fi; source $ctx_temp intshell
 fi
 
@@ -12,8 +12,8 @@ ish_miss_prepare_compile
 ish_miss_prepare_develop
 ish_miss_prepare_operate
 
-# ish_miss_prepare word-dict
 # ish_miss_prepare wubi-dict
+# ish_miss_prepare word-dict
 
 ish_miss_prepare linux-story
 ish_miss_prepare nginx-story
